+++ assets/blocks/djot/index.asset.php
@@ -9,5 +9,5 @@ return [
         'wp-i18n',
         'wp-api-fetch',
     ],
-    'version' => '1.3.1',
+    'version' => '1.4.0',
 ];